+++ src/ai/backend/gateway/etcd.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 '''
 Configuration Schema on etcd
 ----------------------------
@@ -129,6 +131,7 @@ Alias keys are also URL-quoted in the same way.
 '''
 
 import asyncio
+from contextvars import ContextVar
 from collections import defaultdict
 from decimal import Decimal
 import logging
@@ -157,7 +160,12 @@ from ai.backend.common.docker import (
     MIN_KERNELSPEC, MAX_KERNELSPEC,
 )
 from ai.backend.common.logging import BraceStyleAdapter
-from ai.backend.common.types import BinarySize, ResourceSlot
+from ai.backend.common.types import (
+    BinarySize, ResourceSlot,
+    SlotName, SlotTypes,
+    HostPortPair,
+    current_resource_slots,
+)
 from ai.backend.common.exception import UnknownImageReference
 from ai.backend.common.etcd import (
     quote as etcd_quote,
@@ -183,10 +191,16 @@ config_defaults = {
     'config/docker/image/auto_pull': 'digest',
 }
 
+current_vfolder_types: ContextVar[List[str]] = ContextVar('current_vfolder_types')
+
 
 class ConfigServer:
 
-    def __init__(self, app_ctx, etcd_addr, etcd_user, etcd_password, namespace):
+    def __init__(self, app_ctx: Mapping[str, Any],
+                 etcd_addr: HostPortPair,
+                 etcd_user: Optional[str],
+                 etcd_password: Optional[str],
+                 namespace: str) -> None:
         # WARNING: importing etcd3/grpc must be done after forks.
         from ai.backend.common.etcd import AsyncEtcd
         self.context = app_ctx
@@ -211,15 +225,16 @@ class ConfigServer:
                 'A required etcd config is missing.', key)
         return value
 
-    async def register_myself(self, app_config) -> None:
+    async def register_myself(self) -> None:
         instance_id = await get_instance_id()
         manager_info = {
-            'nodes/manager': instance_id,
+            f'nodes/manager/{instance_id}': 'up',
         }
         await self.etcd.put_dict(manager_info)
 
     async def deregister_myself(self) -> None:
-        await self.etcd.delete_prefix('nodes/manager')
+        instance_id = await get_instance_id()
+        await self.etcd.delete_prefix(f'nodes/manager/{instance_id}')
 
     async def update_aliases_from_file(self, file: Path) -> None:
         log.info('Updating image aliases from "{0}"', file)
@@ -300,6 +315,13 @@ class ConfigServer:
             raise UnknownImageReference(reference)
         return await self._parse_image(ref, image_info, reverse_aliases)
 
+    async def forget_image(self, reference: Union[str, ImageRef]):
+        if isinstance(reference, str):
+            ref = await ImageRef.resolve_alias(reference, self.etcd)
+        else:
+            ref = reference
+        await self.etcd.delete_prefix(ref.tag_path)
+
     async def list_images(self):
         known_registries = await get_known_registries(self.etcd)
         reverse_aliases = await self._scan_reverse_aliases()
@@ -311,13 +333,16 @@ class ConfigServer:
             for image, tags in images.items():
                 if image == '':
                     continue
+                if tags == '1':
+                    continue
                 for tag, image_info in tags.items():
                     if tag == '':
                         continue
                     raw_ref = f'{etcd_unquote(registry)}/{etcd_unquote(image)}:{tag}'
                     ref = ImageRef(raw_ref, known_registries)
                     coros.append(self._parse_image(ref, image_info, reverse_aliases))
-        return await asyncio.gather(*coros)
+        result = await asyncio.gather(*coros)
+        return result
 
     async def set_image_resource_limit(self, reference: str, slot_type: str,
                                        value_range: Tuple[Optional[Decimal], Optional[Decimal]]):
@@ -519,13 +544,13 @@ class ConfigServer:
         await asyncio.gather(*coros)
         # TODO: delete images removed from registry?
 
-    async def alias(self, alias: str, target: str):
+    async def alias(self, alias: str, target: str) -> None:
         await self.etcd.put(f'images/_aliases/{etcd_quote(alias)}', target)
 
-    async def dealias(self, alias: str):
+    async def dealias(self, alias: str) -> None:
         await self.etcd.delete(f'images/_aliases/{etcd_quote(alias)}')
 
-    async def update_volumes_from_file(self, file: Path):
+    async def update_volumes_from_file(self, file: Path) -> None:
         log.info('Updating network volumes from "{0}"', file)
         try:
             data = yaml.load(open(file, 'r', encoding='utf-8'))
@@ -541,45 +566,62 @@ class ConfigServer:
             await self.etcd.put_dict(updates)
         log.info('done')
 
-    async def update_resource_slots(self, slot_key_and_units, *,
-                                    clear_existing: bool = True):
+    async def update_resource_slots(
+            self,
+            slot_key_and_units: Mapping[SlotName, SlotTypes]) -> None:
         updates = {}
-        if clear_existing:
-            await self.etcd.delete_prefix('config/resource_slots/')
+        known_slots = await self.get_resource_slots()
         for k, v in slot_key_and_units.items():
-            if k in ('cpu', 'mem'):
-                continue
-            # currently we support only two units
-            # (where count may be fractional)
-            assert v in ('bytes', 'count')
-            updates[f'config/resource_slots/{k}'] = v
-        await self.etcd.put_dict(updates)
+            if k not in known_slots:
+                updates[f'config/resource_slots/{k}'] = v
+        if updates:
+            await self.etcd.put_dict(updates)
 
-    async def update_manager_status(self, status):
+    async def update_manager_status(self, status) -> None:
         await self.etcd.put('manager/status', status.value)
         self.get_manager_status.cache_clear()
 
-    # TODO: refactor using contextvars in Python 3.7 so that the result is cached
-    #       in a per-request basis.
     @aiotools.lru_cache(maxsize=1, expire_after=2.0)
+    async def _get_resource_slots(self):
+        raw_data = await self.etcd.get_prefix_dict('config/resource_slots')
+        return {
+            SlotName(k): SlotTypes(v) for k, v in raw_data.items()
+        }
+
     async def get_resource_slots(self):
         '''
         Returns the system-wide known resource slots and their units.
         '''
-        intrinsic_slots = {'cpu': 'count', 'mem': 'bytes'}
-        configured_slots = await self.etcd.get_prefix_dict('config/resource_slots')
-        return {**intrinsic_slots, **configured_slots}
+        try:
+            ret = current_resource_slots.get()
+        except LookupError:
+            intrinsic_slots = {
+                SlotName('cpu'): SlotTypes('count'),
+                SlotName('mem'): SlotTypes('bytes'),
+            }
+            configured_slots = await self._get_resource_slots()
+            ret = {**intrinsic_slots, **configured_slots}
+            current_resource_slots.set(ret)
+        return ret
 
-    @aiotools.lru_cache(maxsize=1, expire_after=60.0)
+    @aiotools.lru_cache(maxsize=1, expire_after=2.0)
+    async def _get_vfolder_types(self):
+        return await self.etcd.get_prefix_dict('volumes/_types')
+
     async def get_vfolder_types(self):
         '''
         Returns the vfolder types currently set. One of "user" and/or "group".
         If none is specified, "user" type is implicitly assumed.
         '''
-        vf_types = await self.etcd.get_prefix_dict('volumes/_types')
-        if not vf_types:
-            vf_types = {'user': ''}
-        return list(vf_types.keys())
+        try:
+            ret = current_vfolder_types.get()
+        except LookupError:
+            vf_types = await self._get_vfolder_types()
+            if not vf_types:
+                vf_types = {'user': ''}
+            ret = list(vf_types.keys())
+            current_vfolder_types.set(ret)
+        return ret
 
     @aiotools.lru_cache(maxsize=1, expire_after=5.0)
     async def get_manager_nodes_info(self):
@@ -748,7 +790,7 @@ async def delete_config(request: web.Request, params: Any) -> web.Response:
 
 async def init(app: web.Application) -> None:
     if app['pidx'] == 0:
-        await app['config_server'].register_myself(app['config'])
+        await app['config_server'].register_myself()
 
 
 async def shutdown(app: web.Application) -> None: