+++ src/ai/backend/manager/cli/etcd.py
@@ -8,9 +8,15 @@ import sys
 
 import aioredis
 import click
+from tabulate import tabulate
 
 from ai.backend.common.cli import EnumChoice, MinMaxRange
-from ai.backend.common.etcd import AsyncEtcd, ConfigScopes
+from ai.backend.common.docker import ImageRef
+from ai.backend.common.etcd import (
+    AsyncEtcd, ConfigScopes,
+    quote as etcd_quote,
+    unquote as etcd_unquote,
+)
 from ai.backend.common.logging import BraceStyleAdapter
 from ai.backend.gateway.config import redis_config_iv
 from ai.backend.gateway.defs import REDIS_IMAGE_DB
@@ -77,11 +83,15 @@ async def config_ctx(cli_ctx):
 @click.pass_obj
 def put(cli_ctx, key, value, scope):
     '''Put a single key-value pair into the etcd.'''
-    async def _put():
+    async def _impl():
         async with etcd_ctx(cli_ctx) as etcd:
-            await etcd.put(key, value, scope=scope)
+            try:
+                await etcd.put(key, value, scope=scope)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_put())
+        asyncio.run(_impl())
+
 
 
 @cli.command()
@@ -96,25 +106,55 @@ def put_json(cli_ctx, key, file, scope):
     Put a JSON object from FILE to the etcd as flattened key-value pairs
     under the given KEY prefix.
     '''
-    async def _put_json():
-        async with etcd_ctx(cli_ctx) as (loop, etcd):
-            with contextlib.closing(io.BytesIO()) as buf:
-                while True:
-                    part = file.read(65536)
-                    if not part:
-                        break
-                    buf.write(part)
-                value = json.loads(buf.getvalue())
-                value = {f'{key}/{k}': v for k, v in value.items()}
-                await etcd.pit_dict(value, scope=scope)
+    async def _impl():
+        async with etcd_ctx(cli_ctx) as etcd:
+            try:
+                with contextlib.closing(io.BytesIO()) as buf:
+                    while True:
+                        part = file.read(65536)
+                        if not part:
+                            break
+                        buf.write(part)
+                    value = json.loads(buf.getvalue())
+                    value = {f'{key}/{k}': v for k, v in value.items()}
+                    await etcd.put_dict(value, scope=scope)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_put_json())
+        asyncio.run(_impl())
+
+
+@cli.command()
+@click.argument('src_prefix', type=str)
+@click.argument('dst_prefix', type=str)
+@click.option('-s', '--scope', type=EnumChoice(ConfigScopes),
+              default=ConfigScopes.GLOBAL,
+              help='The configuration scope to get/put the subtree. '
+                   'To move between different scopes, use the global scope '
+                   'and specify the per-scope prefixes manually.')
+@click.pass_obj
+def move_subtree(cli_ctx, src_prefix, dst_prefix, scope):
+    '''
+    Move a subtree to another key prefix.
+    '''
+    async def _impl():
+        async with etcd_ctx(cli_ctx) as etcd:
+            try:
+                subtree = await etcd.get_prefix(src_prefix, scope=scope)
+                subtree = {f'{dst_prefix}/{k}': v for k, v in subtree.items()}
+                await etcd.put_dict(subtree, scope=scope)
+                await etcd.delete_prefix(src_prefix, scope=scope)
+            except Exception:
+                log.exception('An error occurred.')
+    with cli_ctx.logger:
+        asyncio.run(_impl())
 
 
 @cli.command()
 @click.argument('key')
 @click.option('--prefix', is_flag=True,
-              help='Get all key-value pairs prefixed with the given key as a JSON form.')
+              help='Get all key-value pairs prefixed with the given key '
+                   'as a JSON form.')
 @click.option('-s', '--scope', type=EnumChoice(ConfigScopes),
               default=ConfigScopes.GLOBAL,
               help='The configuration scope to put the value.')
@@ -123,18 +163,21 @@ def get(cli_ctx, key, prefix, scope):
     '''
     Get the value of a key in the configured etcd namespace.
     '''
-    async def _get():
+    async def _impl():
         async with etcd_ctx(cli_ctx) as etcd:
-            if prefix:
-                data = await etcd.get_prefix(key, scope=scope)
-                print(json.dumps(dict(data), indent=4))
-            else:
-                val = await etcd.get(key, scope=scope)
-                if val is None:
-                    sys.exit(1)
-                print(val)
+            try:
+                if prefix:
+                    data = await etcd.get_prefix(key, scope=scope)
+                    print(json.dumps(dict(data), indent=4))
+                else:
+                    val = await etcd.get(key, scope=scope)
+                    if val is None:
+                        sys.exit(1)
+                    print(val)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_get())
+        asyncio.run(_impl())
 
 
 @cli.command()
@@ -147,36 +190,56 @@ def get(cli_ctx, key, prefix, scope):
 @click.pass_obj
 def delete(cli_ctx, key, prefix, scope):
     '''Delete the key in the configured etcd namespace.'''
-    async def _delete():
+    async def _impl():
         async with etcd_ctx(cli_ctx) as etcd:
-            if prefix:
-                await etcd.delete_prefix(key, scope=scope)
-            else:
-                await etcd.delete(key, scope=scope)
+            try:
+                if prefix:
+                    await etcd.delete_prefix(key, scope=scope)
+                else:
+                    await etcd.delete(key, scope=scope)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_delete())
+        asyncio.run(_impl())
 
 
 @cli.command()
+@click.option('-s', '--short', is_flag=True,
+              help='Show only the image references and digests.')
+@click.option('-i', '--installed', is_flag=True,
+              help='Show only the installed images.')
 @click.pass_obj
-def list_images(cli_ctx):
-    '''List everything about images.'''
-    async def _list_images():
+def list_images(cli_ctx, short, installed):
+    '''List all configured images.'''
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
+            displayed_items = []
             try:
                 items = await config_server.list_images()
-                pprint(items)
+                for item in items:
+                    if installed and not item['installed']:
+                        continue
+                    if short:
+                        img = ImageRef(f"{item['name']}:{item['tag']}",
+                                       item['registry'])
+                        displayed_items.append((img.canonical, item['digest']))
+                    else:
+                        pprint(item)
+                if short:
+                    print(tabulate(displayed_items, tablefmt='plain'))
             except Exception:
                 log.exception('An error occurred.')
-        asyncio.run(_list_images())
+    with cli_ctx.logger:
+        asyncio.run(_impl())
+
 
 
 @cli.command()
 @click.argument('reference')
 @click.pass_obj
 def inspect_image(cli_ctx, reference):
-    '''List everything about images.'''
-    async def _inspect_image():
+    '''Show the details of the given image or alias.'''
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
             try:
                 item = await config_server.inspect_image(reference)
@@ -184,7 +247,27 @@ def inspect_image(cli_ctx, reference):
             except Exception:
                 log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_inspect_image())
+        asyncio.run(_impl())
+
+
+@cli.command()
+@click.argument('reference')
+@click.pass_obj
+def forget_image(cli_ctx, reference):
+    '''
+    Forget (delete) a specific image.
+    NOTE: aliases to the given reference are NOT deleted.
+    '''
+    async def _impl():
+        async with config_ctx(cli_ctx) as config_server:
+            try:
+                await config_server.forget_image(reference)
+                log.info('Done.')
+            except Exception:
+                log.exception('An error occurred.')
+    with cli_ctx.logger:
+        asyncio.run(_impl())
+
 
 
 @cli.command()
@@ -194,14 +277,16 @@ def inspect_image(cli_ctx, reference):
 @click.pass_obj
 def set_image_resource_limit(cli_ctx, reference, slot_type, range_value):
     '''Set the MIN:MAX values of a SLOT_TYPE limit for the given image REFERENCE.'''
-    async def _set_image_resource_limit():
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
             try:
-                await config_server.set_image_resource_limit(reference, slot_type, range_value)
+                await config_server.set_image_resource_limit(
+                    reference, slot_type, range_value)
             except Exception:
                 log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_set_image_resource_limit())
+        asyncio.run(_impl())
+
 
 
 @cli.command()
@@ -213,14 +298,15 @@ def rescan_images(cli_ctx, registry):
 
     Pass the name (usually hostname or "lablup") of the Docker registry configured as REGISTRY.
     '''
-    async def _rescan_images():
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
             try:
                 await config_server.rescan_images(registry)
             except Exception:
                 log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_rescan_images())
+        asyncio.run(_impl())
+
 
 
 @cli.command()
@@ -229,11 +315,15 @@ def rescan_images(cli_ctx, registry):
 @click.pass_obj
 def alias(cli_ctx, alias, target):
     '''Add an image alias from the given alias to the target image reference.'''
-    async def _alias():
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
-            await config_server.alias(alias, target)
+            try:
+                await config_server.alias(alias, target)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_alias())
+        asyncio.run(_impl())
+
 
 
 @cli.command()
@@ -241,8 +331,32 @@ def alias(cli_ctx, alias, target):
 @click.pass_obj
 def dealias(cli_ctx, alias):
     '''Remove an alias.'''
-    async def _dealias():
+    async def _impl():
         async with config_ctx(cli_ctx) as config_server:
-            await config_server.dealias(alias)
+            try:
+                await config_server.dealias(alias)
+            except Exception:
+                log.exception('An error occurred.')
     with cli_ctx.logger:
-        asyncio.run(_dealias())
+        asyncio.run(_impl())
+
+
+@cli.command()
+@click.argument('value')
+@click.pass_obj
+def quote(cli_ctx, value):
+    '''
+    Quote the given string for use as a URL piece in etcd keys.
+    Use this to generate argument inputs for aliases and raw image keys.
+    '''
+    print(etcd_quote(value))
+
+
+@cli.command()
+@click.argument('value')
+@click.pass_obj
+def unquote(cli_ctx, value):
+    '''
+    Unquote the given string used as a URL piece in etcd keys.
+    '''
+    print(etcd_unquote(value))