+++ tests/conftest.py
@@ -12,7 +12,6 @@ import signal
 import subprocess
 import tempfile
 
-import aiodocker
 import aiohttp
 from aiohttp import web
 import aiohttp_cors
@@ -24,8 +23,8 @@ import psycopg2 as pg
 import pytest
 
 from ai.backend.common.argparse import host_port_pair
-from ai.backend.common.types import HostPortPair
-from ai.backend.gateway.config import load as load_config
+from ai.backend.gateway.config import load as load_config, redis_config_iv
+from ai.backend.gateway.etcd import ConfigServer
 from ai.backend.gateway.server import (
     gw_init, gw_shutdown,
     exception_middleware, api_middleware,
@@ -37,14 +36,16 @@ from ai.backend.manager.models import agents, kernels, keypairs, vfolders
 here = Path(__file__).parent
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope='session', autouse=True)
 def test_id():
     return secrets.token_hex(12)
 
 
-@pytest.fixture(scope='session')
+@pytest.fixture(scope='session', autouse=True)
 def test_ns(test_id):
-    return f'testing-ns-{test_id}'
+    ret = f'testing-ns-{test_id}'
+    os.environ['BACKEND_NAMESPACE'] = ret
+    return ret
 
 
 @pytest.fixture(scope='session')
@@ -53,62 +54,100 @@ def test_db(test_id):
 
 
 @pytest.fixture(scope='session')
-def folder_mount(test_id):
-    return Path(f'/tmp/backend.ai/vfolders-{test_id}')
+def vfolder_mount(test_id):
+    ret = Path(f'/tmp/backend.ai-testing/vfolders-{test_id}')
+    ret.mkdir(parents=True, exist_ok=True)
+    yield ret
+    shutil.rmtree(ret.parent)
 
 
 @pytest.fixture(scope='session')
-def folder_fsprefix(test_id):
+def vfolder_fsprefix(test_id):
     # NOTE: the prefix must NOT start with "/"
     return Path('fsprefix/inner/')
 
 
 @pytest.fixture(scope='session')
-def folder_host():
+def vfolder_host():
     return 'local'
 
 
-@pytest.fixture(scope='session', autouse=True)
-def prepare_and_cleanup_databases(request, test_ns, test_db,
-                                  folder_mount, folder_host, folder_fsprefix):
-    os.environ['BACKEND_NAMESPACE'] = test_ns
-    os.environ['BACKEND_DB_NAME'] = test_db
-
-    # Clear and reset etcd namespace using CLI functions.
+@pytest.fixture(scope='session')
+def test_config(test_id, test_db):
     cfg = load_config()
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'volumes/_mount', str(folder_mount)])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'volumes/_fsprefix', str(folder_fsprefix)])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'volumes/_default_host', str(folder_host)])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'config/docker/registry/index.docker.io',
-                     'https://registry-1.docker.io'])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'config/redis/addr', '127.0.0.1:8110'])
-    # Add fake plugin settings.
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'config/plugins/cloudia/base_url', '127.0.0.1:8090'])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'config/plugins/cloudia/user', 'fake-cloudia-user@lablup.com'])
-    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'put',
-                     'config/plugins/cloudia/password', 'fake-password'])
+    cfg['db']['name'] = test_db
+    # In normal setups, this is read from etcd.
+    cfg['redis'] = redis_config_iv.check({
+        'addr': {'host': '127.0.0.1', 'port': '6379'},
+    })
+    return cfg
 
-    def finalize_etcd():
-        subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'delete',
-                         '', '--prefix'])
-    request.addfinalizer(finalize_etcd)
 
+@pytest.fixture(scope='session')
+def etcd_fixture(test_id, test_config, vfolder_mount, vfolder_fsprefix, vfolder_host):
+    # Clear and reset etcd namespace using CLI functions.
+    with tempfile.NamedTemporaryFile(mode='w', suffix='.etcd.json') as f:
+        etcd_fixture = {
+            'volumes': {
+                '_mount': str(vfolder_mount),
+                '_fsprefix': str(vfolder_fsprefix),
+                '_default_host': str(vfolder_host),
+            },
+            'nodes': {
+            },
+            'config': {
+                'docker': {
+                    'registry': {
+                        'index.docker.io': 'https://registry-1.docker.io',
+                    },
+                },
+                'redis': {
+                    'addr': '127.0.0.1:6379',
+                },
+                'plugins': {
+                    'cloudia': {
+                        'base_url': '127.0.0.1:8090',
+                        'user': 'fake-cloudia-user@lablup.com',
+                        'password': 'fake-password',
+                    }
+                }
+            },
+        }
+        json.dump(etcd_fixture, f)
+        f.flush()
+        subprocess.call([
+            'python', '-m', 'ai.backend.manager.cli',
+            'etcd', 'put-json', '', f.name,
+        ])
+    yield
+    subprocess.call(['python', '-m', 'ai.backend.manager.cli', 'etcd', 'delete',
+                     '', '--prefix'])
+
+
+@pytest.fixture
+async def config_server(app, etcd_fixture):
+    server = ConfigServer(
+        app,
+        app['config']['etcd']['addr'],
+        app['config']['etcd']['user'],
+        app['config']['etcd']['password'],
+        app['config']['etcd']['namespace'],
+    )
+    yield server
+
+
+@pytest.fixture(scope='session')
+def database(request, test_config, test_db):
     # Create database using low-level psycopg2 API.
-    db_addr = cfg['db']['addr']
-    db_user = cfg['db']['user']
-    db_pass = cfg['db']['password']
+    db_addr = test_config['db']['addr']
+    db_user = test_config['db']['user']
+    db_pass = test_config['db']['password']
+    # Temporarily use "testing" dbname until we create our own db.
     if db_pass:
         # TODO: escape/urlquote db_pass
-        db_url = f'postgresql://{db_user}:{db_pass}@{db_addr}'
+        db_url = f'postgresql://{db_user}:{db_pass}@{db_addr}/testing'
     else:
-        db_url = f'postgresql://{db_user}@{db_addr}'
+        db_url = f'postgresql://{db_user}@{db_addr}/testing'
     conn = pg.connect(db_url)
     conn.set_isolation_level(pg.extensions.ISOLATION_LEVEL_AUTOCOMMIT)
     cur = conn.cursor()
@@ -130,7 +169,7 @@ def prepare_and_cleanup_databases(request, test_ns, test_db,
     request.addfinalizer(finalize_db)
 
     # Load the database schema using CLI function.
-    alembic_url = db_url + '/' + test_db
+    alembic_url = f'postgresql://{db_user}:{db_pass}@{db_addr}/{test_db}'
     with tempfile.NamedTemporaryFile(mode='w', encoding='utf8') as alembic_cfg:
         alembic_sample_cfg = here / '..' / 'alembic.ini.sample'
         alembic_cfg_data = alembic_sample_cfg.read_text()
@@ -212,30 +251,12 @@ class Client:
 
 
 @pytest.fixture
-async def app(event_loop, test_ns, test_db, unused_tcp_port_factory):
-    """
-    For tests that do not require actual server running.
-    """
+async def app(event_loop, test_config, unused_tcp_port_factory):
     app = web.Application(middlewares=[
         exception_middleware,
         api_middleware,
     ])
-    app['config'] = load_config()
-
-    # Override settings for testing.
-    app['config']['db']['name'] = test_db
-    app['config']['etcd']['namespace'] = test_ns
-    app['config']['manager']['num-proc'] = 2
-    app['config']['manager']['heartbeat-timeout'] = 10.0
-    app['config']['manager']['service-addr'] = HostPortPair(
-        '127.0.0.1', unused_tcp_port_factory())
-    # import ssl
-    # app['config'].ssl_cert = here / 'sample-ssl-cert' / 'sample.crt'
-    # app['config'].ssl_key = here / 'sample-ssl-cert' / 'sample.key'
-    # app['sslctx'] = ssl.SSLContext(ssl.PROTOCOL_SSLv23)
-    # app['sslctx'].load_cert_chain(str(app['config'].ssl_cert),
-    #                            str(app['config'].ssl_key))
-
+    app['config'] = test_config
     app['pidx'] = 0
     return app
 
@@ -295,7 +316,7 @@ async def monitor_keypair(event_loop, app):
 
 
 @pytest.fixture
-def get_headers(app, default_keypair, prepare_docker_images):
+def get_headers(app, default_keypair):
     def create_header(method, url, req_bytes, ctype='application/json',
                       hash_type='sha256', api_version='v4.20181215',
                       keypair=default_keypair):
@@ -338,7 +359,7 @@ def get_headers(app, default_keypair, prepare_docker_images):
 
 @pytest.fixture
 async def create_app_and_client(request, test_id, test_ns,
-                                event_loop, app,
+                                event_loop, app, database,
                                 default_keypair, user_keypair):
     client = None
     runner = None
@@ -389,7 +410,7 @@ async def create_app_and_client(request, test_id, test_ns,
         await runner.setup()
         site = web.TCPSite(
             runner,
-            app['config']['manager']['service-addr'].host,
+            str(app['config']['manager']['service-addr'].host),
             app['config']['manager']['service-addr'].port,
             ssl_context=app.get('sslctx'),
             **server_params,
@@ -476,30 +497,6 @@ async def create_app_and_client(request, test_id, test_ns,
     if extra_proc:
         os.kill(extra_proc.pid, signal.SIGINT)
         extra_proc.join()
-
-
-@pytest.fixture(scope='session')
-def prepare_docker_images():
-    _loop = asyncio.new_event_loop()
-
-    async def pull():
-        docker = aiodocker.Docker()
-        images_to_pull = [
-            'lablup/kernel-lua:5.3-alpine',
-        ]
-        for img in images_to_pull:
-            try:
-                await docker.images.inspect(img)
-            except aiodocker.exceptions.DockerError as e:
-                assert e.status == 404
-                print(f'Pulling image "{img}" for testing...')
-                await docker.pull(img)
-        await docker.close()
-
-    try:
-        _loop.run_until_complete(pull())
-    finally:
-        _loop.close()
 
 
 @pytest.fixture